% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Account-class}
\alias{Account}
\alias{Account-class}
\title{A Reference Class to represent a bank account.}
\description{
A Reference Class to represent a bank account.
}
\section{Fields}{

\describe{
\item{\code{balance1}}{A length-one numeric vector}

\item{\code{balance2}}{A length-one numeric vector}
}}
\section{Methods}{

\describe{
\item{\code{withdraw1(x)}}{Withdraw1 money from account. Allows overdrafts}

\item{\code{withdraw2(x)}}{Withdraw2 money from account. Allows overdrafts}
}}

