% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-delete-records.R
\name{redcap_delete_records}
\alias{redcap_delete_records}
\title{Delete one ore more records from REDCap}
\usage{
redcap_delete_records(
  redcap_uri,
  token,
  record_ids,
  verbose = TRUE,
  config_options = NULL
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.  Required.}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{record_ids}{A single record id or an array of ids to delete.
Make sure these ids exist in REDCap.}

\item{verbose}{A boolean value indicating if \code{message}s should be printed
to the R console during the operation.  The verbose output might contain
sensitive information (\emph{e.g.} PHI), so turn this off if the output might
be visible somewhere public. Optional.}

\item{config_options}{A list of options to pass to \code{POST} method in the
\code{httr} package.  See the details in \code{\link[=redcap_read_oneshot]{redcap_read_oneshot()}}. Optional.}
}
\value{
The number of records deleted.
}
\description{
Delete existing records by their ID from REDCap.
}
\examples{
\dontrun{
uri   <- "https://bbmc.ouhsc.edu/redcap/api/"
token <- "9A81268476645C4E5F03428B8AC3AA7B"
redcap_delete_records(redcap_uri = redcap_uri,
                     token = token,
                     record_ids = record_ids)
}
}
\author{
Jonathan Mang
}
