% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-instrument-repeating.R
\name{redcap_instrument_repeating}
\alias{redcap_instrument_repeating}
\title{Enumerate repeating instruments and events}
\usage{
redcap_instrument_repeating(
  redcap_uri,
  token,
  verbose = TRUE,
  config_options = NULL,
  handle_httr = NULL
)
}
\arguments{
\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{verbose}{A boolean value indicating if \code{message}s should be printed
to the R console during the operation.  The verbose output might contain
sensitive information (\emph{e.g.} PHI), so turn this off if the output might
be visible somewhere public. Optional.}

\item{config_options}{A list of options passed to \code{\link[httr:POST]{httr::POST()}}.
See details at \code{\link[httr:httr_options]{httr::httr_options()}}. Optional.}

\item{handle_httr}{The value passed to the \code{handle} parameter of
\code{\link[httr:POST]{httr::POST()}}.
This is useful for only unconventional authentication approaches.  It
should be \code{NULL} for most institutions.  Optional.}
}
\value{
Currently, a list is returned with the following elements,
\itemize{
\item \code{data}: A \code{\link[tibble:tibble]{tibble::tibble()}} where each row represents one repeating instrument-event
combination in the REDCap project.
\item \code{success}: A boolean value indicating if the operation was apparently
successful.
\item \code{status_code}: The
\href{https://en.wikipedia.org/wiki/List_of_HTTP_status_codes}{http status code}
of the operation.
\item \code{outcome_message}: A human readable string indicating the operation's
outcome.
\item \code{elapsed_seconds}: The duration of the function.
\item \code{raw_text}: If an operation is NOT successful, the text returned by
REDCap.  If an operation is successful, the \code{raw_text} is returned as an
empty string to save RAM.
}
}
\description{
This method allows you to export a list of the repeated instruments
and repeating events for a project. This includes their unique
instrument name as seen in the second column of the Data
Dictionary, as well as each repeating instrument's corresponding
custom repeating instrument label. For longitudinal projects, the
unique event name is also returned for each repeating instrument.
Additionally, repeating events are returned as separate items,
in which the instrument name will be blank/null to indicate that
it is a repeating event (rather than a repeating instrument).
(Copied from "Export Repeating Instruments and Events" method of
REDCap API documentation, v.16.0.4)
}
\examples{
\dontrun{
uri                 <- "https://redcap-dev-2.ouhsc.edu/redcap/api/"

# Repeating events project
token_1  <- "FEF7A22B52E6B9942AFF7A28C426C871" # "repeating-instruments" test project
REDCapR::redcap_instrument_repeating(redcap_uri=uri, token=token_1)$data

# Classic project (without repeating instruments) throws an error
token_2  <- "9A068C425B1341D69E83064A2D273A70" # "simple" test project
# REDCapR::redcap_instrument_repeating(redcap_uri=uri, token=token_2)$data
}
}
\references{
The official documentation can be found on the 'API Help Page'
and 'API Examples' pages on the REDCap wiki (\emph{i.e.},
https://community.projectredcap.org/articles/456/api-documentation.html and
https://community.projectredcap.org/articles/462/api-examples.html).
If you do not have an account for the wiki, please ask your campus REDCap
administrator to send you the static material.
}
\author{
Ezra Porter
}
