% Generated by roxygen2 (4.0.2): do not edit by hand
\name{redcap_metadata_read}
\alias{redcap_metadata_read}
\title{Export the metadata of a REDCap project.}
\usage{
redcap_metadata_read(redcap_uri, token, forms = NULL, forms_collapsed = "",
  fields = NULL, fields_collapsed = "", verbose = TRUE,
  cert_location = NULL)
}
\arguments{
  \item{redcap_uri}{The URI (uniform resource identifier)
  of the REDCap project.  Required.}

  \item{token}{The user-specific string that serves as the
  password for a project.  Required.}

  \item{forms}{An array, where each element corresponds to
  the REDCap form of the desired fields.  Optional.}

  \item{forms_collapsed}{A single string, where the desired
  forms are separated by commas.  Optional.}

  \item{fields}{An array, where each element corresponds to
  a desired project field.  Optional.}

  \item{fields_collapsed}{A single string, where the
  desired field names are separated by commas.  Optional.}

  \item{verbose}{A boolean value indicating if
  \code{message}s should be printed to the R console during
  the operation.  The verbose output might contain
  sensitive information (\emph{e.g.} PHI), so turn this off
  if the output might be visible somewhere public.
  Optional.}

  \item{cert_location}{If present, this string should point
  to the location of the cert files required for SSL
  verification.  If the value is missing or NULL, the
  server's identity will be verified using a recent CA
  bundle from the \href{http://curl.haxx.se}{cURL website}.
  See the details below. Optional.}
}
\value{
Currently, a list is returned with the following elements,
\enumerate{ \item \code{data}: An R \code{data.frame} of
the desired records and columns.  \item \code{success}: A
boolean value indicating if the operation was apparently
successful.  \item \code{status_codes}: A collection of
\href{http://en.wikipedia.org/wiki/List_of_HTTP_status_codes}{http
status codes}, separated by semicolons.  There is one code
for each batch attempted.  \item \code{outcome_messages}: A
collection of human readable strings indicating the
operations' semicolons.  There is one code for each batch
attempted.  In an unsuccessful operation, it should contain
diagnostic information.  \item \code{forms_collapsed}: The
desired records IDs, collapsed into a single string,
separated by commas.  \item \code{fields_collapsed}: The
desired field names, collapsed into a single string,
separated by commas.  \item \code{elapsed_seconds}: The
duration of the function. }
}
\description{
Export the metadata (as a data dictionary) of a REDCap
project as a \code{data.frame}. Each row in the data
dictionary corresponds to one field in the project's
dataset.
}
\details{
Specifically, it internally uses multiple calls to
\code{\link{redcap_read_oneshot}} to select and return
data. Initially, only primary key is queried through the
REDCap API.  The long list is then subsetted into
partitions, whose sizes are determined by the
\code{batch_size} parameter.  REDCap is then queried for
all variables of the subset's subjects.  This is repeated
for each subset, before returning a unified
\code{data.frame}.

The function allows a delay between calls, which allows the
server to attend to other users' requests.
}
\examples{
\dontrun{
library(REDCapR) #Load the package into the current R session.
}
}
\author{
Will Beasley
}
\references{
The official documentation can be found on the REDCap wiki
(\url{https://iwg.devguard.com/trac/redcap/wiki/ApiDocumentation}).
A user account is required to access the wiki, which
typically is granted only to REDCap administrators.
}

