% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{redcap_project}
\alias{redcap_project}
\title{A \code{Reference Class} to make later calls to REDCap more convenient.}
\description{
This \code{Reference Class} represents a REDCap project.
Once some values are set that are specific to a REDCap
project (such as the URI and token), later calls are less
verbose (such as reading and writing data).
}
\examples{
library(REDCapR) #Load the package into the current R session.
uri <- "https://bbmc.ouhsc.edu/redcap/api/"
token <- "9A81268476645C4E5F03428B8AC3AA7B"
project <- redcap_project$new(redcap_uri=uri, token=token)
project$read()
}

