% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REDCapR-package.R
\docType{package}
\name{REDCapR}
\alias{REDCapR}
\alias{REDCapR-package}
\title{REDCapR}
\description{
Thanks to Funders, including \href{https://perf-data.hrsa.gov/mchb/DGISReports/Abstract/AbstractDetails.aspx?Source=TVIS&GrantNo=D89MC23154&FY=2012}{HRSA/ACF D89MC23154}

\emph{OUHSC CCAN Independent Evaluation of the State of Oklahoma Competitive Maternal, Infant, and Early Childhood Home Visiting
(\href{http://mchb.hrsa.gov/programs/homevisiting/}{MIECHV}) Project.}, which
evaluates MIECHV expansion and enhancement of Evidence-based Home Visitation programs in four Oklahoma counties.
}

