% Generated by roxygen2 (4.0.2): do not edit by hand
\name{retrieve_token}
\alias{retrieve_token}
\alias{retrieve_token_mssql}
\title{Read a token from a (non-REDCap) database.}
\usage{
retrieve_token_mssql(dsn_name, project_name, channel = NULL,
  schema_name = "[Redcap]", procedure_name = "[prcToken]",
  variable_name_project = "@RedcapProjectName", field_name_token = "Token")
}
\arguments{
  \item{dsn_name}{---.  The default is 100.}

  \item{project_name}{---. The default is 0.5 seconds.}

  \item{channel}{---.  Required.}

  \item{schema_name}{---.  Required.}

  \item{procedure_name}{---.  Optional.}

  \item{variable_name_project}{---.  Optional.}

  \item{field_name_token}{---.  Optional.}
}
\value{
The token, which is a 32 character string.
}
\description{
From an external perspective, this function is similar to
\code{\link{redcap_read_oneshot}}.  The internals differ in
that \code{redcap_read} retrieves subsets of the data, and
then combines them before returning (among other objects) a
single \code{data.frame}.  This function can be more
appropriate than \code{\link{redcap_read_oneshot}} when
returning large datasets that could tie up the server.
}
\details{
Specifically, it internally uses multiple calls to
\code{\link{redcap_read_oneshot}} to select and return
data. Initially, only primary key is queried through the
REDCap API.  The long list is then subsetted into
partitions, whose sizes are determined by the
\code{batch_size} parameter.  REDCap is then queried for
all variables of the subset's subjects.  This is repeated
for each subset, before returning a unified
\code{data.frame}.

The function allows a delay between calls, which allows the
server to attend to other users' requests.
}
\examples{
\dontrun{
library(REDCapR) #Load the package into the current R session.
dsn_name <- "TokenSecurity"
project_name <- "DiabetesSurveyProject2"
token <- retrieve_token(dsn_name=dsn, project_name=project)
}
}
\author{
Will Beasley
}
\references{
The official documentation can be found on the REDCap wiki
(\url{https://iwg.devguard.com/trac/redcap/wiki/ApiDocumentation}).
Also see the `API Examples' page on the REDCap wiki
(\url{https://iwg.devguard.com/trac/redcap/wiki/ApiExamples}).
A user account is required to access the wiki, which
typically is granted only to REDCap administrators. If you
do not

The official \href{http://curl.haxx.se}{cURL site}
discusses the process of using SSL to verify the server
being connected to.
}

