% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{constant}
\alias{constant}
\title{Collection of REDCap-specific constants}
\usage{
constant(name)
}
\arguments{
\item{name}{Name of constant.  Required character.}
}
\value{
The constant's value.  Currently all are single integers, but that could be expanded in the future.
}
\description{
Collection of constants defined by the REDCap developers.
}
\details{
The current constants relate to the 'complete' variable at the end of each form.
\itemize{
\item \code{form_incomplete}: 0L (\emph{i.e.}, an integer)
\item \code{form_unverified}: 1L
\item \code{form_complete}: 2L
}

To add more, please for and edit the
\href{https://github.com/OuhscBbmc/REDCapR/blob/master/R/constant.R}{constant.R}
on GitHub and submit a pull request.  For instructions, please see
\href{https://help.github.com/articles/editing-files-in-another-user-s-repository/}{Editing files in another user's repository}
in the GitHub documentation.
}
\examples{

constant("form_incomplete")  # Returns 0L
constant("form_unverified")  # Returns 1L
constant("form_complete")    # Returns 2L

\dontrun{
# The following line returns an error:
#     Assertion on 'name' failed: Must be a subset of
#     {'form_complete','form_incomplete','form_unverified'},
#     but is {'bad-name'}.

constant("bad-name")    # Returns an error
}
}
\author{
Will Beasley
}
