% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-download-file-oneshot.R
\name{redcap_download_file_oneshot}
\alias{redcap_download_file_oneshot}
\title{Download a file from a REDCap project record}
\usage{
redcap_download_file_oneshot(file_name = NULL, directory = NULL,
  overwrite = FALSE, redcap_uri, token, record, field, event = "",
  verbose = TRUE, config_options = NULL)
}
\arguments{
\item{file_name}{The name of the file where the downloaded file is saved.
If empty the original name of the file will be used and saved in the default directory.  Optional.}

\item{directory}{The directory where the file is saved. By default current directory. Optional}

\item{overwrite}{Boolean value indicating if existing files should be overwritten. Optional}

\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap project.  Required.}

\item{token}{The user-specific string that serves as the password for a project.  Required.}

\item{record}{The record ID where the file is to be imported. Required}

\item{field}{The name of the field where the file is saved in REDCap. Required}

\item{event}{The name of the event where the file is saved in REDCap. Optional}

\item{verbose}{A boolean value indicating if \code{message}s should be printed to the R console during the operation.  Optional.}

\item{config_options}{A list of options to pass to \code{\link[httr:POST]{httr::POST()}} method in the 'httr' package.  See the details below. Optional.}
}
\value{
Currently, a list is returned with the following elements,
\itemize{
\item \code{success}: A boolean value indicating if the operation was apparently successful.
\item \code{status_code}: The \href{http://en.wikipedia.org/wiki/List_of_HTTP_status_codes}{http status code} of the operation.
\item \code{outcome_message}: A human readable string indicating the operation's outcome.
\item \code{records_affected_count}: The number of records inserted or updated.
\item \code{affected_ids}: The subject IDs of the inserted or updated records.
\item \code{elapsed_seconds}: The duration of the function.
\item \code{raw_text}: If an operation is NOT successful, the text returned by REDCap.  If an operation is successful, the \code{raw_text} is returned as an empty string to save RAM.
\item \code{file_name}: The name of the file persisted to disk. This is useful if the name stored in REDCap is used (which is the default).
}
}
\description{
This function uses REDCap's API to download a file.
}
\details{
Currently, the function doesn't modify any variable types to conform to REDCap's supported variables.  See \code{\link[=validate_for_write]{validate_for_write()}} for a helper function that checks for some common important conflicts.
}
\examples{
\dontrun{
uri     <- "https://bbmc.ouhsc.edu/redcap/api/"
token   <- "D70F9ACD1EDD6F151C6EA78683944E98" #pid=213
record  <- 1
field   <- "mugshot"
# event <- "" # only for longitudinal events

result_1 <- REDCapR::redcap_download_file_oneshot(
  record        = record,
  field         = field,
  redcap_uri    = uri,
  token         = token
)
base::unlink("mugshot-1.jpg")

(full_name <- base::tempfile(pattern="mugshot", fileext=".jpg"))
result_2   <- REDCapR::redcap_download_file_oneshot(
  file_name     = full_name,
  record        = record,
  field         = field,
  redcap_uri    = uri,
  token         = token
)
base::unlink(full_name)

(relative_name <- "ssss.jpg")
result_3 <- REDCapR::redcap_download_file_oneshot(
  file_name    = relative_name,
  record       = record,
  field        = field,
  redcap_uri   = uri,
  token        = token
)
base::unlink(relative_name)
}
}
\references{
The official documentation can be found on the 'API Help Page' and 'API Examples' pages
on the REDCap wiki (\emph{i.e.}, https://community.projectredcap.org/articles/456/api-documentation.html and
https://community.projectredcap.org/articles/462/api-examples.html). If you do not have an account
for the wiki, please ask your campus REDCap administrator to send you the static material.
}
\author{
Will Beasley, John J. Aponte
}
