% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-metadata-coltypes.R
\name{redcap_metadata_coltypes}
\alias{redcap_metadata_coltypes}
\title{Suggests a col_type for each field in a REDCap project}
\usage{
redcap_metadata_coltypes(
  redcap_uri,
  token,
  http_response_encoding = "UTF-8",
  locale = readr::default_locale(),
  verbose = FALSE,
  config_options = NULL
)
}
\arguments{
\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{http_response_encoding}{The encoding value passed to
\code{\link[httr:content]{httr::content()}}.  Defaults to 'UTF-8'.}

\item{locale}{a \code{\link[readr:locale]{readr::locale()}} object to specify preferences like
number, date, and time formats.  This object is passed to
\code{\link[readr:read_delim]{readr::read_csv()}}.  Defaults to \code{\link[readr:locale]{readr::default_locale()}}.}

\item{verbose}{A boolean value indicating if \code{message}s should be printed
to the R console during the operation.  The verbose output might contain
sensitive information (\emph{e.g.} PHI), so turn this off if the output might
be visible somewhere public. Optional.}

\item{config_options}{A list of options to pass to \code{POST} method in the
\code{httr} package.  See the details below. Optional.}
}
\value{
A \code{\link[readr:cols]{readr::cols()}} object is returned, which can be
passed to \code{\link[=redcap_read]{redcap_read()}} or \code{\link[=redcap_read_oneshot]{redcap_read_oneshot()}}.

Additionally objected is printed to the console, see the Details below.
}
\description{
This function inspects a REDCap project to
determine a \code{\link[readr:cols]{readr::cols()}} object that is compatible with the
the project's current definition.
}
\details{
\code{redcap_metadata_coltypes()} returns a \code{\link[readr:cols]{readr::cols()}} object in two ways.
First, a literal object is returned that can be passed to
\code{\link[=redcap_read]{redcap_read()}} or \code{\link[=redcap_read_oneshot]{redcap_read_oneshot()}}.

Second, the function acts as a code generator.
It prints text to the console so that it can be copied
and pasted into an R file.  This is useful to (a) document what
fields and data types are expected, and (b) adjust those fields and
data types if the defaults can be customized for your needs.
For instance, you may choose to exclude some variables or tweak a
data type (\emph{e.g.}, changing a patient's height from an integer to
a double).

The full list of configuration options accepted by the \code{httr} package is
viewable by executing \code{\link[httr:httr_options]{httr::httr_options()}}.  The \code{httr} package and
documentation is available at https://cran.r-project.org/package=httr.
}
\examples{
\dontrun{
uri      <- "https://bbmc.ouhsc.edu/redcap/api/"

# A simple project
token      <- "9A81268476645C4E5F03428B8AC3AA7B" # 153
col_types  <- redcap_metadata_coltypes(uri, token)
redcap_read_oneshot(uri, token, col_types = col_types)$data

# A longitudinal project
token      <- "0434F0E9CF53ED0587847AB6E51DE762" # 212
col_types  <- redcap_metadata_coltypes(uri, token)
redcap_read_oneshot(uri, token, col_types = col_types)$data

# A repeating instruments project
token      <- "56F43A10D01D6578A46393394D76D88F" # 2603
col_types  <- redcap_metadata_coltypes(uri, token)
redcap_read_oneshot(uri, token, col_types = col_types)$data

# A project with every field type and validation type.
#   Notice It throws a warning that some fields use a comma for a decimal,
#   while other fields use a period/dot as a decimal
token      <- "8F5313CAA266789F560D79EFCEE2E2F1" # 2634 - Validation Types
col_types  <- redcap_metadata_coltypes(uri, token)
redcap_read_oneshot(uri, token, col_types = col_types)$data
}
}
\references{
The official documentation can be found on the 'API Help Page'
and 'API Examples' pages on the REDCap wiki (\emph{i.e.},
https://community.projectredcap.org/articles/456/api-documentation.html and
https://community.projectredcap.org/articles/462/api-examples.html).
If you do not have an account for the wiki, please ask your campus REDCap
administrator to send you the static material.
}
\author{
Will Beasley
}
