% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-project.R
\name{redcap_project}
\alias{redcap_project}
\alias{RedcapProject}
\title{A Reference Class to make later calls to REDCap more convenient}
\description{
This \verb{Reference Class} represents a REDCap project.
Once some values are set that are specific to a REDCap project
(such as the URI and token), later calls are less verbose
(such as reading and writing data).
}
\examples{
uri     <- "https://bbmc.ouhsc.edu/redcap/api/"
token   <- "D70F9ACD1EDD6F151C6EA78683944E98"
\dontrun{
project <- REDCapR::redcap_project$new(redcap_uri=uri, token=token)
ds_all  <- project$read()

# Demonstrate how repeated calls are more concise when the token and
#   url aren't always passed.
ds_skinny <- project$read(fields=c("record_id", "sex", "height"))$data

ids_of_males    <- ds_skinny$record_id[ds_skinny$sex==1]
ids_of_shorties <- ds_skinny$record_id[ds_skinny$height < 40]

ds_males        <- project$read(records=ids_of_males, batch_size=2)$data
ds_shorties     <- project$read(records=ids_of_shorties)$data

#Switch the Genders
sex_original   <- ds_skinny$sex
ds_skinny$sex  <- (1 - ds_skinny$sex)
project$write(ds_skinny)

#Switch the Genders back
ds_skinny$sex <- sex_original
project$write(ds_skinny)
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{redcap_uri}}{The URI (uniform resource identifier) of the
REDCap project.  Required.}

\item{\code{token}}{token The user-specific string that serves as the
password for a project.  Required.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{redcap_uri}}{The URI (uniform resource identifier) of the
REDCap project.  Required.}

\item{\code{token}}{token The user-specific string that serves as the
password for a project.  Required.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RedcapProject$new()}}
\item \href{#method-createProject}{\code{RedcapProject$createProject()}}
\item \href{#method-clone}{\code{RedcapProject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedcapProject$new(redcap_uri, redcap_token = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createProject"></a>}}
\if{latex}{\out{\hypertarget{method-createProject}{}}}
\subsection{Method \code{createProject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedcapProject$createProject(
  redcap_super_token = NULL,
  title,
  notes = "",
  purpose = 0,
  purpose_other = "",
  is_longitudinal = FALSE,
  is_survey = FALSE,
  is_autonumbering = FALSE,
  verbose = TRUE,
  config_options = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedcapProject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
