% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-project-info-read.R
\name{redcap_project_info_read}
\alias{redcap_project_info_read}
\title{Export project information.}
\usage{
redcap_project_info_read(
  redcap_uri,
  token,
  http_response_encoding = "UTF-8",
  locale = readr::default_locale(),
  verbose = TRUE,
  config_options = NULL
)
}
\arguments{
\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{http_response_encoding}{The encoding value passed to
\code{\link[httr:content]{httr::content()}}.  Defaults to 'UTF-8'.}

\item{locale}{a \code{\link[readr:locale]{readr::locale()}} object to specify preferences like
number, date, and time formats.  This object is passed to
\code{\link[readr:read_delim]{readr::read_csv()}}.  Defaults to \code{\link[readr:locale]{readr::default_locale()}}.}

\item{verbose}{A boolean value indicating if \code{message}s should be printed
to the R console during the operation.  The verbose output might contain
sensitive information (\emph{e.g.} PHI), so turn this off if the output might
be visible somewhere public. Optional.}

\item{config_options}{A list of options to pass to \code{POST} method in the
\code{httr} package.}
}
\value{
Currently, a list is returned with the following elements:
\itemize{
\item \code{data}: An R \code{\link[tibble:tibble]{tibble::tibble()}} of all data access groups of the project.
\item \code{success}: A boolean value indicating if the operation was apparently
successful.
\item \code{status_codes}: A collection of
\href{https://en.wikipedia.org/wiki/List_of_HTTP_status_codes}{http status codes},
separated by semicolons.  There is one code for each batch attempted.
\item \code{outcome_messages}: A collection of human readable strings indicating the
operations' semicolons.  There is one code for each batch attempted.  In an
unsuccessful operation, it should contain diagnostic information.
\item \code{elapsed_seconds}: The duration of the function.
}
}
\description{
This function exports some of the basic attributes of a given
REDCap project, such as the project's title, if it is longitudinal,
if surveys are enabled, the time the project was created and moved to
production.  Returns a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\examples{
\dontrun{
uri                  <- "https://bbmc.ouhsc.edu/redcap/api/"
token_simple         <- "9A81268476645C4E5F03428B8AC3AA7B"
token_longitudinal   <- "0434F0E9CF53ED0587847AB6E51DE762"

d1 <- REDCapR::redcap_project_info_read(uri, token_simple      )$data
View(d1)

d2 <- REDCapR::redcap_project_info_read(uri, token_longitudinal)$data
View(d2)


# Stack all the projects on top of each other in a (nested) tibble
#   (starting from a csv of REDCapR test projects).
d_all <-
  system.file("misc/example.credentials", package = "REDCapR") |>
  readr::read_csv(
    comment     = "#",
    col_select  = c(redcap_uri, token),
    col_types   = readr::cols(.default = readr::col_character())
  ) |>
  dplyr::filter(32L == nchar(token)) |>
  purrr::pmap_dfr(REDCapR::redcap_project_info_read)

# Inspect values stored on the server.
View(d_all$data)

# Inspect everything returned, including values like the http status code.
View(d_all)
}
}
\references{
The official documentation can be found on the 'API Help Page'
and 'API Examples' pages on the REDCap wiki (\emph{i.e.},
https://community.projectredcap.org/articles/456/api-documentation.html
and
https://community.projectredcap.org/articles/462/api-examples.html).
If you do not have an account for the wiki, please ask your campus REDCap
administrator to send you the static material.
}
\author{
Will Beasley, Stephan Kadauke
}
