% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-variables.R
\name{redcap_variables}
\alias{redcap_variables}
\title{Enumerate the exported variables}
\usage{
redcap_variables(
  redcap_uri,
  token,
  delimiter = ",",
  verbose = TRUE,
  config_options = NULL,
  handle_httr = NULL
)
}
\arguments{
\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{delimiter}{A single-character value passed to
the \code{delim} parameter of \code{\link[readr:read_delim]{readr::read_delim()}}.
Options include:
\enumerate{
\item \verb{,} (a comma, the default),
\item \verb{;} (a semi-colon),
\item \code{|} (a pipe),
\item \code{^} (a caret), or
\item \verb{\\t} (a tab).
}}

\item{verbose}{A boolean value indicating if \code{message}s should be printed
to the R console during the operation.  The verbose output might contain
sensitive information (\emph{e.g.} PHI), so turn this off if the output might
be visible somewhere public. Optional.}

\item{config_options}{A list of options passed to \code{\link[httr:POST]{httr::POST()}}.
See details at \code{\link[httr:httr_options]{httr::httr_options()}}. Optional.}

\item{handle_httr}{The value passed to the \code{handle} parameter of
\code{\link[httr:POST]{httr::POST()}}.
This is useful for only unconventional authentication approaches.  It
should be \code{NULL} for most institutions.  Optional.}
}
\value{
Currently, a list is returned with the following elements,
\itemize{
\item \code{data}: A \code{\link[tibble:tibble]{tibble::tibble()}} where each row represents one column
in the REDCap dataset.
\item \code{success}: A boolean value indicating if the operation was apparently
successful.
\item \code{status_code}: The
\href{https://en.wikipedia.org/wiki/List_of_HTTP_status_codes}{http status code}
of the operation.
\item \code{outcome_message}: A human readable string indicating the operation's
outcome.
\item \code{elapsed_seconds}: The duration of the function.
\item \code{raw_text}: If an operation is NOT successful, the text returned by
REDCap.  If an operation is successful, the \code{raw_text} is returned as an
empty string to save RAM.
}
}
\description{
This function calls the 'exportFieldNames' function of the
REDCap API.
}
\details{
As of REDCap version 6.14.2, three variable types are \emph{not} returned in
this call: calculated, file, and descriptive.  All variables returned are
writable/uploadable.
}
\examples{
\dontrun{
uri         <- "https://redcap-dev-2.ouhsc.edu/redcap/api/"
token       <- "9A068C425B1341D69E83064A2D273A70"

ds_variable <- REDCapR::redcap_variables(redcap_uri=uri, token=token)$data
}
}
\references{
The official documentation can be found on the 'API Help Page'
and 'API Examples' pages on the REDCap wiki (\emph{i.e.},
\url{https://redcap.vumc.org/community/post.php?id=456} and
\url{https://redcap.vumc.org/community/post.php?id=462} ).
If you do not have an account for the wiki, please ask your campus REDCap
administrator to send you the static material.
}
\author{
Will Beasley
}
