% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-version.R
\name{redcap_version}
\alias{redcap_version}
\title{Determine version of REDCap instance}
\usage{
redcap_version(
  redcap_uri,
  token,
  verbose = TRUE,
  config_options = NULL,
  handle_httr = NULL
)
}
\arguments{
\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{token}{The user-specific string that serves as the password for
a project.  Required.}

\item{verbose}{A boolean value indicating if \code{message}s should be printed
to the R console during the operation.  The verbose output might contain
sensitive information (\emph{e.g.} PHI), so turn this off if the output might
be visible somewhere public. Optional.}

\item{config_options}{A list of options passed to \code{\link[httr:POST]{httr::POST()}}.
See details at \code{\link[httr:httr_options]{httr::httr_options()}}. Optional.}

\item{handle_httr}{The value passed to the \code{handle} parameter of
\code{\link[httr:POST]{httr::POST()}}.
This is useful for only unconventional authentication approaches.  It
should be \code{NULL} for most institutions.  Optional.}
}
\value{
a \link[utils:packageDescription]{utils::packageDescription}
}
\description{
This function uses REDCap's API to query its version.
}
\details{
If the API call is unsuccessful, a value of
\code{base::package_version("0.0.0")} will be returned.
This ensures that a the function will always return an object of class
\link[base:numeric_version]{base::numeric_version}.  It guarantees the value can always be used in
\code{\link[utils:compareVersion]{utils::compareVersion()}}.
}
\examples{
uri      <- "https://redcap-dev-2.ouhsc.edu/redcap/api/"
token    <- "9A068C425B1341D69E83064A2D273A70"
REDCapR::redcap_version(redcap_uri = uri, token = token)
}
